+++ internal/app/microservices/chat/usecase.go
@@ -6,10 +6,10 @@ import (
 )
 
 type ChatUsecase interface {
-	CreateChat(req *models.ChatCreateReq, userID int) (*models.ChatResponse, *errors.Error)
-	GetChatById(chatID int, userID int) (*models.ChatResponse, *errors.Error)
-	GetUserChats(userID int) ([]*models.ChatResponse, *errors.Error)
-	CreateMessage(req *models.CreateMessageReq, userID int) (*models.MessageResp, *errors.Error)
+	CreateChat(req *models.ChatCreateReq, userID uint64) (*models.ChatResponse, *errors.Error)
+	GetChatById(chatID uint64, userID uint64) (*models.ChatResponse, *errors.Error)
+	GetUserChats(userID uint64) ([]*models.ChatResponse, *errors.Error)
+	CreateMessage(req *models.CreateMessageReq, userID uint64) (*models.MessageResp, *errors.Error)
 	GetLastNMessages(req *models.GetLastNMessagesReq) ([]*models.MessageResp, *errors.Error)
 	GetNMessagesBefore(req *models.GetNMessagesBeforeReq) ([]*models.MessageResp, *errors.Error)
 }
